+++ packages/cli-core/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @msw-dev-tool/cli-core
 
+## 0.2.4
+
+### Patch Changes
+
+- Updated dependencies [8d996fa]
+  - @msw-dev-tool/core@1.6.0
+
+## 0.2.3
+
+### Patch Changes
+
+- Updated dependencies [67de2dd]
+- Updated dependencies [bda8f30]
+  - @msw-dev-tool/core@1.5.1
+
 ## 0.2.2
 
 ### Patch Changes